pyproject.toml
[tool.poetry]
name = "bank-mini-project"
version = "0.1.0"
description = ""
authors = ["wkdsh21 <wkdsh21@naver.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-restframework = "^0.0.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
django-stubs = "^5.1.1"
psycopg2 = "^2.9.10"
tokens = "^0.0.3"
dj-rest-auth = "^7.0.0"
types-jwt = "^0.1.3"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework-stubs = "^3.15.1"
pillow = "^11.0.0"
pandas = "^2.2.3"
matplotlib = "^3.9.3"


[tool.poetry.group.dev.dependencies]
drf-yasg = "^1.21.8"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.base"



